Removing dead code from modules
There are some modules with code that is not being executed
because PYAOSCX is a requirement for this collection.

This CL removes the code that would be executed in the case of
PYAOSCX is not present. Also some unused legacy module utils are
removed.

(cherry picked from commit 6466d222b382038083a999f7a33444bb0e8b23ab)