Fix MAC Ansible Module
When MACs are gathered from switch, multiple requests are performed
and affects the performance when a high number of MACs and VLANs
are configured. For each VLAN a GET operation is performed and also
for each MAC address.

This CL modifies this behavior and gets all the information with
just one GET operation.

CR 261730

(cherry picked from commit 6dc91335f55a4bb34b307320688db7765775a0bf)