Fix aoscx_facts module
Facts module is failing on 6000 and 6100 with an "401 Unauthorized"
message; this is because session is closed while querying the
version of the firmware.

This commit adds a reconnect step if the session closes in that
step; also some error handling is included.

CR 268602

(cherry picked from commit d7207318746e2e55731838bdbb968a4bfde4ebc1)