Fix aoscx_facts module
When stressing the device, if 1000 VLANS are added to the switch,
module facts is not able to display the physical interfaces.
If VLANS are deleted from the config, facts module displays the
information correctly again.

This commit fixes performance and module works in that scenario.

CR 268603

(cherry picked from commit 15cc5cad89a64887b9ba6db5219fd8d105b17ffa)