Fix for vrf attach and idempotency
When trying to delete the VRF attached to an interface, the
interface becomes disabled and when the VRF attach is changed to
another, the existing configuration of the interface remains, so
also, its idempotence fails and does not return a change.
This commit fixes these flaws and add a change for idempotency
when adding IPv4 or IPv6.

AOSANSI 268665
AOSANSI 268663
AOSANSI 268664

(cherry picked from commit c22baae65477bb1a2c112b99eaa8a4705d8f042e)