Fix to Keep the  previous VRF attach
When an interface has a VRF attached and additional configuration
is applied, the current VRF changes to the default VRF.
This commit allows to keep the previous vrf, and only when it is
removed the vrf is changed to the default VRF value. Also, is added
a change in the documentation.

(cherry picked from commit 6879233927899fbf8b2661cbb729400e4b30a117)