Fix for delete of VRF
When the VRF is removed, the interfaces attached to the VRF keep
config. This commit adds changes to imitate the CLI behavior, when
the VRF is removed, all the configuration of the interfaces with
this VRF attached is removed.

AOSANSI 268666

(cherry picked from commit d58886d95c30b418f18b6e905f1753364f42ed62)